added showing info about class

+++ src/main/java/ru/mudan/domain/repositories/ClassRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.mudan.domain.entity.ClassEntity;
 
 @Repository
-public interface ClassRepository extends JpaRepository<ClassEntity, Integer> {
+public interface ClassRepository extends JpaRepository<ClassEntity, Long> {
 }
